Rename file to have a more proper naming convention
